+++ stevedore/dispatch.py
@@ -1,4 +1,8 @@
+import logging
+
 from .enabled import EnabledExtensionManager
+
+LOG = logging.getLogger(__name__)
 
 
 class DispatchExtensionManager(EnabledExtensionManager):
@@ -72,8 +76,14 @@ class NameDispatchExtensionManager(DispatchExtensionManager):
     different inputs to different extensions and can predict the name
     of the extensions before calling them.
 
+    The check_func argument should return a boolean, with ``True``
+    indicating that the extension should be loaded and made available
+    and ``False`` indicating that the extension should be ignored.
+
     :param namespace: The namespace for the entry points.
     :type namespace: str
+    :param check_func: Function to determine which extensions to load.
+    :type check_func: callable
     :param invoke_on_load: Boolean controlling whether to invoke the
         object returned by the entry point after the driver is loaded.
     :type invoke_on_load: bool
@@ -85,6 +95,7 @@ class NameDispatchExtensionManager(DispatchExtensionManager):
         the object returned by the entry point. Only used if invoke_on_load
         is True.
     :type invoke_kwds: dict
+
     """
 
     def __init__(self, namespace, check_func, invoke_on_load=False,
@@ -95,7 +106,7 @@ class NameDispatchExtensionManager(DispatchExtensionManager):
             invoke_on_load=invoke_on_load,
             invoke_args=invoke_args,
             invoke_kwds=invoke_kwds,
-            )
+        )
         self.by_name = dict((e.name, e) for e in self.extensions)
 
     def map(self, names, func, *args, **kwds):
@@ -120,6 +131,10 @@ class NameDispatchExtensionManager(DispatchExtensionManager):
         """
         response = []
         for name in names:
-            e = self.by_name[name]
-            self._invoke_one_plugin(response.append, func, e, args, kwds)
+            try:
+                e = self.by_name[name]
+            except KeyError:
+                LOG.debug('Missing extension %r being ignored', name)
+            else:
+                self._invoke_one_plugin(response.append, func, e, args, kwds)
         return response