+++ stevedore/tests/test_named.py
@@ -10,7 +10,7 @@ def test_named():
         invoke_on_load=True,
         invoke_args=('a',),
         invoke_kwds={'b': 'B'},
-        )
+    )
     actual = em.names()
     assert actual == ['t1']
 
@@ -32,6 +32,27 @@ def test_enabled_before_load():
             invoke_on_load=True,
             invoke_args=('a',),
             invoke_kwds={'b': 'B'},
-            )
+        )
         actual = em.names()
         assert actual == []
+
+
+def test_extensions_listed_in_name_order():
+    # Since we don't know the "natural" order of the extensions, run
+    # the test both ways: if the sorting is broken, one of them will
+    # fail
+    em = named.NamedExtensionManager(
+        'stevedore.test.extension',
+        names=['t1', 't2'],
+        name_order=True
+    )
+    actual = em.names()
+    assert actual == ['t1', 't2']
+
+    em = named.NamedExtensionManager(
+        'stevedore.test.extension',
+        names=['t2', 't1'],
+        name_order=True
+    )
+    actual = em.names()
+    assert actual == ['t2', 't1']